requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.9.0
python-multipart==0.0.6
langchain-openai==0.2.0
python-dotenv==1.0.0
openai==1.51.0
streamlit